Create a base test class for facility-level tests
